requirements.txt
langchain==0.3.9
langchain_chroma==0.1.4
langchain_community==0.3.8
langchain_openai==0.2.10
streamlit==1.40.2
pypdf==5.1.0
python-dotenv==1.0.1
